move stories inline with components
