+++ src/components/app/StatusesSummary.stories.tsx
@@ -2,12 +2,9 @@ import { TestStepResultStatus } from '@cucumber/messages'
 import { Story } from '@ladle/react'
 import React from 'react'
 
-import { components } from '../../src'
-import { CucumberReact } from '../components'
-import { IStatusesSummaryProps } from '../components/app'
-import { makeEmptyScenarioCountsByStatus } from '../countScenariosByStatuses'
-
-const { StatusesSummary } = components.app
+import { makeEmptyScenarioCountsByStatus } from '../../countScenariosByStatuses'
+import { CucumberReact } from '../CucumberReact'
+import { IStatusesSummaryProps, StatusesSummary } from './StatusesSummary'
 
 export default {
   title: 'App/StatusesSummary',
