+++ src/components/customise/Classes.stories.tsx
@@ -1,14 +1,12 @@
-import './custom-classes.scss'
+import './custom-classes.stories.scss'
 
 import * as messages from '@cucumber/messages'
 import { Story } from '@ladle/react'
 import React from 'react'
 
-import { components } from '..'
-import { CustomRenderingSupport } from '../components/customise'
-
-const { DocString } = components.gherkin
-const { CucumberReact } = components
+import { CucumberReact } from '../CucumberReact'
+import { DocString } from '../gherkin'
+import { CustomRenderingSupport } from './customRendering'
 
 export default {
   title: 'Customisation/Classes',
