+++ src/components/customise/Themes.stories.tsx
@@ -2,13 +2,11 @@ import * as messages from '@cucumber/messages'
 import { Story } from '@ladle/react'
 import React from 'react'
 
-import attachments from '../../acceptance/attachments/attachments.feature'
-import rules from '../../acceptance/rules/rules.feature'
-import { components } from '..'
-import { IncludedTheme } from '../components/customise'
-
-const { EnvelopesWrapper, GherkinDocumentList } = components.app
-const { CucumberReact } = components
+import attachments from '../../../acceptance/attachments/attachments.feature'
+import rules from '../../../acceptance/rules/rules.feature'
+import { EnvelopesWrapper, GherkinDocumentList } from '../app'
+import { CucumberReact } from '../CucumberReact'
+import { IncludedTheme } from './theming'
 
 export default {
   title: 'Customisation/Themes',
