+++ src/components/gherkin/GherkinDocument.stories.tsx
@@ -2,24 +2,22 @@ import * as messages from '@cucumber/messages'
 import { Story } from '@ladle/react'
 import React from 'react'
 
-import attachments from '../../acceptance/attachments/attachments.feature'
-import dataTables from '../../acceptance/data-tables/data-tables.feature'
-import examplesTables from '../../acceptance/examples-tables/examples-tables.feature'
-import hooks from '../../acceptance/hooks/hooks.feature'
-import markdown from '../../acceptance/markdown/markdown.feature.md'
-import minimal from '../../acceptance/minimal/minimal.feature'
-import parameterTypes from '../../acceptance/parameter-types/parameter-types.feature'
-import pending from '../../acceptance/pending/pending.feature'
-import retry from '../../acceptance/retry/retry.feature'
-import rules from '../../acceptance/rules/rules.feature'
-import skipped from '../../acceptance/skipped/skipped.feature'
-import stacktTraces from '../../acceptance/stack-traces/stack-traces.feature'
-import undefinedEnvelopes from '../../acceptance/undefined/undefined.feature'
-import unknownParameterTypes from '../../acceptance/unknown-parameter-type/unknown-parameter-type.feature'
-import { components } from '../../src'
-import { CucumberReact } from '../components'
-
-const { EnvelopesWrapper, GherkinDocumentList } = components.app
+import attachments from '../../../acceptance/attachments/attachments.feature'
+import dataTables from '../../../acceptance/data-tables/data-tables.feature'
+import examplesTables from '../../../acceptance/examples-tables/examples-tables.feature'
+import hooks from '../../../acceptance/hooks/hooks.feature'
+import markdown from '../../../acceptance/markdown/markdown.feature.md'
+import minimal from '../../../acceptance/minimal/minimal.feature'
+import parameterTypes from '../../../acceptance/parameter-types/parameter-types.feature'
+import pending from '../../../acceptance/pending/pending.feature'
+import retry from '../../../acceptance/retry/retry.feature'
+import rules from '../../../acceptance/rules/rules.feature'
+import skipped from '../../../acceptance/skipped/skipped.feature'
+import stacktTraces from '../../../acceptance/stack-traces/stack-traces.feature'
+import undefinedEnvelopes from '../../../acceptance/undefined/undefined.feature'
+import unknownParameterTypes from '../../../acceptance/unknown-parameter-type/unknown-parameter-type.feature'
+import { EnvelopesWrapper, GherkinDocumentList } from '../app'
+import { CucumberReact } from '../CucumberReact'
 
 export default {
   title: 'Gherkin/GherkinDocument',
