+++ tsconfig.build.json
@@ -14,6 +14,6 @@
   ],
   "exclude": [
     "src/**/*.spec.*",
-    "src/stories"
+    "src/**/*.stories.*",
   ]
 }
